🐛 fix bugs

+++ backend/travelapp/src/main/java/com/sms/travelapp/repository/PinRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PinRepository extends JpaRepository<Pin,Integer> {
+public interface PinRepository extends JpaRepository<Pin,Long> {
 
 
 }
